feat: add strings utils from suite to utils package
