+++ packages/utils/src/capitalizeFirstLetter.ts
@@ -1,0 +1,1 @@
+export const capitalizeFirstLetter = (s: string) => s.charAt(0).toUpperCase() + s.slice(1);
