+++ packages/utils/src/countBytesInString.ts
@@ -1,0 +1,1 @@
+export const countBytesInString = (input: string) => encodeURI(input).split(/%..|./).length - 1;
