+++ packages/utils/src/getNumberFromPxString.ts
@@ -1,0 +1,1 @@
+export const getNumberFromPxString = (size: string): number => parseInt(size.replace('px', ''), 10);
