+++ packages/utils/tests/capitalizeFirstLetter.test.ts
@@ -1,0 +1,6 @@
+import { capitalizeFirstLetter } from '../src/capitalizeFirstLetter';
+
+it('capitalizeFirstLetter', () => {
+    expect(capitalizeFirstLetter('god')).toBe('God');
+    expect(capitalizeFirstLetter('dog')).toBe('Dog');
+});
