+++ packages/utils/tests/getNumberFromPxString.test.ts
@@ -1,0 +1,6 @@
+import { getNumberFromPxString } from '../src/getNumberFromPxString';
+
+it('getNumberFromPxString', () => {
+    expect(getNumberFromPxString('1px')).toBe(1);
+    expect(getNumberFromPxString('1')).toBe(1);
+});
